pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "newsletter-pod"
version = "0.1.0"
description = "Private Apple Podcasts daily digest from RSS newsletters"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.3.0",
  "PyJWT[crypto]>=2.9.0",
  "feedparser>=6.0.11",
  "requests>=2.32.0",
  "PyYAML>=6.0.1",
  "python-dateutil>=2.9.0",
  "tzdata>=2024.1",
  "google-cloud-firestore>=2.17.0",
  "google-cloud-storage>=2.17.0",
  "google-cloud-secret-manager>=2.20.0",
  "google-cloud-tasks>=2.16.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-mock>=3.14.0",
  "httpx>=0.28.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.setuptools.packages.find]
include = ["newsletter_pod*"]
